+++ front-end/src/components/Login/EmailInput.jsx
@@ -4,14 +4,7 @@ function EmailInput({ value, onChange }) {
   return (
     <div className="form-group">
       <label>Email</label>
-      <input
-        type="text"
-        id="inputEmail"
-        value={value}
-        onChange={onChange}
-        placeholder="이메일"
-        required
-      />
+      <input type="text" id="inputEmail" value={value} onChange={onChange} placeholder="이메일" required />
     </div>
   );
 }
