+++ front-end/src/components/Login/PwInput.jsx
@@ -4,14 +4,7 @@ function PwInput({ value, onChange }) {
   return (
     <div className="form-group">
       <label>Password</label>
-      <input
-        type="password"
-        id="inputPw"
-        value={value}
-        onChange={onChange}
-        placeholder="비밀번호"
-        required
-      />
+      <input type="password" id="inputPw" value={value} onChange={onChange} placeholder="비밀번호" required />
     </div>
   );
 }
