+++ front-end/src/components/Main/Banner/index.styled.js
