+++ front-end/src/components/Main/Cards/index.styled.js
