+++ front-end/src/components/Main/Category/index.styled.js
