+++ front-end/src/components/Main/Explore/index.styled.js
