+++ front-end/src/components/Main/Text/index.jsx
