+++ front-end/src/components/Main/Text/index.styled.js
