+++ front-end/src/pages/Join/index.jsx
@@ -1,4 +1,4 @@
-import RegisterForm from "./RegisterForm";
+import RegisterForm from "../../components/Join/RegisterForm";
 import "./index.styled.css";
 import { useNavigate } from "react-router-dom";
 import { useRecoilValue } from "recoil";
